Move the files around
