+++ src/Camelotia.Presentation.Tests/Services/ByteConverterTests.cs
@@ -1,8 +1,8 @@
-using Camelotia.Presentation.Extensions;
+using Camelotia.Services.Extensions;
 using FluentAssertions;
 using Xunit;
 
-namespace Camelotia.Presentation.Tests
+namespace Camelotia.Presentation.Tests.Services
 {
     public sealed class ByteConverterTests
     {
