+++ src/Camelotia.Presentation.Tests/Services/LocalFileSystemProviderTests.cs
@@ -8,7 +8,7 @@ using Camelotia.Services.Providers;
 using FluentAssertions;
 using Xunit;
 
-namespace Camelotia.Presentation.Tests
+namespace Camelotia.Presentation.Tests.Services
 {
     public sealed class LocalFileSystemProviderTests
     {
