+++ src/Camelotia.Presentation.Tests/Services/VkontakteProviderTests.cs
@@ -1,4 +1,5 @@
-using System;
+﻿using System;
+using System.IO;
 using Akavache;
 using Camelotia.Services.Models;
 using Camelotia.Services.Providers;
@@ -6,32 +7,32 @@ using FluentAssertions;
 using NSubstitute;
 using Xunit;
 
-namespace Camelotia.Presentation.Tests
+namespace Camelotia.Presentation.Tests.Services
 {
-    public sealed class GoogleDriveProviderTests
+    public sealed class VkontakteFileSystemProviderTests
     {
         private readonly IBlobCache _blobCache = Substitute.For<IBlobCache>();
         private readonly ProviderModel _model = new ProviderModel
         {
             Id = Guid.NewGuid(),
-            Created = DateTime.Now,
-            Type = "Google Drive"
+            Type = "Vkontakte",
+            Created = DateTime.Now
         };
 
         [Fact]
         public void VerifyDefaultPropertyValues()
         {
-            var provider = new GoogleDriveProvider(_model, _blobCache);
-            provider.InitialPath.Should().Be("/");
+            var provider = new VkDocsProvider(_model, _blobCache);
+            provider.InitialPath.Should().Be(Path.DirectorySeparatorChar.ToString());
 
             provider.CanCreateFolder.Should().BeFalse();
             provider.Created.Should().Be(_model.Created);
-            provider.Name.Should().Be("Google Drive");
+            provider.Name.Should().Be("Vkontakte");
             provider.Id.Should().Be(_model.Id);
 
-            provider.SupportsDirectAuth.Should().BeFalse();
+            provider.SupportsDirectAuth.Should().BeTrue();
             provider.SupportsHostAuth.Should().BeFalse();
-            provider.SupportsOAuth.Should().BeTrue();
+            provider.SupportsOAuth.Should().BeFalse();
         }
     }
 }
