+++ src/Camelotia.Presentation/ViewModels/ProviderViewModel.cs
@@ -8,8 +8,8 @@ using System.Reactive.Concurrency;
 using System.Reactive.Linq;
 using System.Reactive.Disposables;
 using System.Reactive.Threading.Tasks;
-using Camelotia.Presentation.Extensions;
 using Camelotia.Presentation.Interfaces;
+using Camelotia.Services.Extensions;
 using Camelotia.Services.Interfaces;
 using Camelotia.Services.Models;
 using ReactiveUI.Fody.Helpers;
